start unit tests, modify controllers

+++ MonoRec/Repositories/PatientRepository.cs
@@ -9,10 +9,12 @@ namespace MonoRec.Repositories
     public class PatientRepository : IPatientRepository
     {
         MonoRecDbContext _db;
+        // IEnumerable<Patient> patients; (look at this for testing)
 
         public PatientRepository(MonoRecDbContext db)
         {
             _db = db;
+            // patients = db.Patients; (look at this for testing)
         }
 
         public IEnumerable<Patient> GetAllPatients()
@@ -23,7 +25,7 @@ namespace MonoRec.Repositories
 
         public Patient GetPatient(int patId)
         {
-            var patient = _db.Patients.Where(patient => patient.PatientId == patId).First();
+            var patient = _db.Patients.FirstOrDefault(patient => patient.PatientId == patId);
             return patient;
 
             // create conditional to check if patient is in database, if not throw error
@@ -39,6 +41,10 @@ namespace MonoRec.Repositories
 
         public IEnumerable<Doctor> GetAllDoctorsByPatient(int patId)
         {
+            var patient = _db.Patients.FirstOrDefault(patient => patient.PatientId == patId);
+
+            if (patient == null) return null;
+
             var doctors = _db.Doctors.ToList();
             var doctorsPatients = _db.DoctorsPatients.ToList();
 
@@ -49,28 +55,41 @@ namespace MonoRec.Repositories
                             select doctor;
 
             return innerJoin.ToList();
-
+            
+            // create conditional to check if patient is in database, if not throw error - still works, but change response to error anyways
+            // if no doctors are available for patient, then this currently returns empty list (which I think is fine to keep)
         }
 
         public Doctor AddNewDoctorForPatient(int patId, int docId)
         {
+            var doctorToAdd = _db.Doctors.FirstOrDefault(doctor => doctor.DoctorId == docId);
+            var patientToModify = _db.Patients.FirstOrDefault(patient => patient.PatientId == patId);
+
+            if (doctorToAdd == null || patientToModify == null) return null;
+
             var newDoctorPatient = new DoctorPatient(docId, patId);
             _db.DoctorsPatients.Add(newDoctorPatient);
             _db.SaveChanges();
 
-            var doctorToAdd = _db.Doctors.Where(doctor => doctor.DoctorId == docId).First();
             return doctorToAdd;
+
+            // create conditional to check if patient is in database, if not throw error
         }
 
         public Doctor DeleteDoctorForPatient(int patId, int docId)
         {
+            var doctorToDelete = _db.Doctors.FirstOrDefault(doctor => doctor.DoctorId == docId);
+            var patientToModify = _db.Patients.FirstOrDefault(patient => patient.PatientId == patId);
+
+            if (doctorToDelete == null || patientToModify == null) return null;
+
             var doctorPatientToDelete = _db.DoctorsPatients.Where(doctorPatient => doctorPatient.DoctorId == docId && doctorPatient.PatientId == patId).First();
             _db.DoctorsPatients.Remove(doctorPatientToDelete);
             _db.SaveChanges();
 
-            var doctorToDelete = _db.Doctors.Where(doctor => doctor.DoctorId == docId).First();
             return doctorToDelete;
 
+            // create conditional to check if patient is in database, if not throw error
         }
     }
 }
